# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CheckoutRequest": {
        "additionalProperties": false,
        "properties": {
          "plan_id": {
            "format": "uuid",
            "title": "Plan Id",
            "type": "string"
          }
        },
        "required": [
          "plan_id"
        ],
        "title": "CheckoutRequest",
        "type": "object"
      },
      "CheckoutResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "payment_id": {
            "format": "uuid",
            "title": "Payment Id",
            "type": "string"
          },
          "price": {
            "$ref": "#/components/schemas/MoneyResponse"
          },
          "provider_available": {
            "title": "Provider Available",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "payment_id",
          "status",
          "price",
          "provider_available",
          "detail"
        ],
        "title": "CheckoutResponse",
        "type": "object"
      },
...
      "ExportCheckoutResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "payment_id": {
            "format": "uuid",
            "title": "Payment Id",
            "type": "string"
          },
          "provider_available": {
            "title": "Provider Available",
            "type": "boolean"
          },
          "purchase": {
            "$ref": "#/components/schemas/ExportPurchaseResponse"
          }
        },
        "required": [
          "purchase",
          "payment_id",
          "provider_available",
          "detail"
        ],
        "title": "ExportCheckoutResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Zylora V2 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/billing/subscription/checkouts": {
      "post": {
        "operationId": "checkout_api_v1_billing_subscription_checkouts_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Checkout",
        "tags": [
          "commerce"
        ]
      }
    },
...
    "/api/v1/website-exports/{purchase_id}": {
      "get": {
        "operationId": "export_status_api_v1_website_exports__purchase_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "purchase_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Purchase Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportPurchaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Status",
        "tags": [
          "commerce"
        ]
      }
    },
    "/api/v1/website-exports/{purchase_id}/checkout": {
      "post": {
        "operationId": "export_checkout_api_v1_website_exports__purchase_id__checkout_post",
        "parameters": [
          {
            "in": "path",
            "name": "purchase_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Purchase Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportCheckoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Checkout",
        "tags": [
          "commerce"
        ]
      }
    },
    "/api/v1/website-exports/{purchase_id}/download": {
      "get": {
        "operationId": "download_export_api_v1_website_exports__purchase_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "purchase_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Purchase Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Export",
        "tags": [
          "commerce"
        ]
      }
    },
    "/api/v1/website-exports/{purchase_id}/generate": {
      "post": {
        "operationId": "generate_export_api_v1_website_exports__purchase_id__generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "purchase_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Purchase Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportGenerationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Export",
        "tags": [
          "commerce"
        ]
      }
    },
...
    "/api/v1/websites/{website_id}/exports": {
      "post": {
        "operationId": "create_export_purchase_api_v1_websites__website_id__exports_post",
        "parameters": [
          {
            "in": "path",
            "name": "website_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Website Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportPurchaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Export Purchase",
        "tags": [
          "commerce"
        ]
      }
    },
...
    "/api/v1/websites/{website_id}/publish": {
      "post": {
        "operationId": "publish_api_v1_websites__website_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "website_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Website Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishCommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish",
        "tags": [
          "commerce"
        ]
      }
    },
    "/api/v1/websites/{website_id}/publish-evaluation": {
      "get": {
        "operationId": "publish_evaluation_api_v1_websites__website_id__publish_evaluation_get",
        "parameters": [
          {
            "in": "path",
            "name": "website_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Website Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "domain_type",
            "required": false,
            "schema": {
              "default": "ZYLORA_SUBDOMAIN",
              "pattern": "^(ZYLORA_SUBDOMAIN|CUSTOM)$",
              "title": "Domain Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishEvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Evaluation",
        "tags": [
          "commerce"
        ]
      }
    },
...
    "/api/v1/websites/{website_id}/transfers": {
      "post": {
        "operationId": "transfer_api_v1_websites__website_id__transfers_post",
        "parameters": [
          {
            "in": "path",
            "name": "website_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Website Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OwnershipTransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnershipTransferResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transfer",
        "tags": [
          "commerce"
        ]
      }
    },
...
  }
}